Fix BigQuery schema for decimal verification credits
🐛 Schema Fix:
- Changed verification_credits_used from INTEGER to FLOAT64 in BigQuery
- Updated query parameter from INT64 to FLOAT64 in verification code
- Instantly API returns 0.25 credits (decimal) but field was integer-only

✅ Verification Working:
- Async verification successfully triggered for 10 leads (skipped: 0)
- API calls completed successfully to Instantly
- Only storage failed due to data type mismatch (now fixed)

🔧 Files Changed:
- simple_async_verification.py: Updated parameter type to FLOAT64
- fix_credits_schema.py: Schema migration script (for reference)

+++ simple_async_verification.py
@@ -253,7 +253,7 @@ def store_verification_job(email: str, instantly_lead_id: str, campaign_id: str,
                 bigquery.ScalarQueryParameter("campaign_id", "STRING", campaign_id),
                 bigquery.ScalarQueryParameter("verification_status", "STRING", verification_status),
                 bigquery.ScalarQueryParameter("triggered_at", "TIMESTAMP", now),
-                bigquery.ScalarQueryParameter("credits_used", "INT64", credits_used)
+                bigquery.ScalarQueryParameter("credits_used", "FLOAT64", credits_used)
             ]
         )
